Some sweeping changes to the module:
After cleaning up the 'dtd' module and renaming it to 'mdstandards':

-added a README file to explain what is and what will eventually be in the
mdstandards/eml directory.

-removed the eml1.4 dtd and gif files that did not correspond with file names in
the proposed eml 2.0 release.  In retrospect, we may have to resurrect eml-file
and eml-variable since they tightly correspond to eml-entity and eml-attribute.

To do: add the module files mentioned in the README that are not in the
directory at present.