This commit introduced major structural changes to the access schema document, but the concepts essentially remained the same.  I've changed the structure in response to our discussions about being conservative with named complexTypes:
- Got rid of the 'ACL' complexType, but put it within the 'acl' element as an
unamed complexType.  It merely defines a sequence within that element now.

- Got rid of the 'Resource' complex type.  Again, I placed this into the
'resource' element as an unnamed complexType containing a sequence of elements.
In doing so, the 'order' attribute for the old Resource type was moved to the
'resource' element.

- The PermissionType complexType has been renamed to the 'AccessRule'
complexType to reflect the dtd more cosely, and to better describe the allow and
deny elements.