After discussions with Matt and our meeting last week, I've made changes to eml-access.[dtd,xsd]:
1)Simplified the the xsd file by moving the named simpleTypes into their
respective xs:attributes (permission, order), which eliminated the need for the
Permission and Order named simpleTypes.

2)Removed the public attribute, knowing that we'll need to change morpho and
metacat implementations, but that we need the most up to date changes for our
April EML meeting.

3)Imported eml-resource per Matt, anticipating changes to resource where an
'Identifier' type is created with an attribute of 'system' that will be used to
allow multiple identifiers for a document that are used in system specific
context (i.e. metacat, OBFS, etc...).  I also changed the dtd to include an
entity of resource in order to use the identifier element defined within
resource.

Now that we've included un-named simpleTypes within elements and attributes,
we'll need to provide support for that in the eml-documentation.xsl file.