Chnages to support the new XML Schema PR mechanisms for indicating extension and restriction.  Also, a variety of formatting cleanup.
There are still some minor problems with the way in which derived
types are handled. Under certain specific situations (see LitCitation in
eml-resource) the type from which a complexType is derived is listed twice
in the description ofthe content model.

Need to go through and prune the apply-template select statements because
I think we call many more things than are allowed by the XML Schema model
(for example, elements can't contain other elements in the PR, but we
allow it in the documentation, and shouldn't).