Added a slightly modified version of the stylesheet that Dan Higgins wrote for stripping all of the annotation tags out of the XSD files.  This is useful to get a sense of the structure of the XSD files without the documentation getting in the way.
Also modified the build.xml file so that there is now a new ant target that
allows running the above stylesheet.  Running:

     ant stripdocs

causes ant to call XSLT and process all of the xsd files in the current
directory and create a stripped version in a directory of choice (by
default called "stripped").