1) removed all of the xml declarations from the beginning of all of the dtds
2) added header comments to all of the dtds to include license, cvs info, etc.

3) for eml-coverage.dtd, I changed the content model for dsployo and dspolyx
after our discussion today.  Since the fgdc standard calls for:

((grngpoin, grngpoin, grngpoin, grngpoin+) | gring)

which seems incorrect (defining a polygon with at least 4 lat/lon points, rather
than 3 lat/lon points) (or a gring set), and since a polygon with 3 or 4 that
are colinear don't define a very useful polygon, we decided to change the
content model to:

(grngpoin+ | gring)

and would state in the documentation that polygons should be defined with 3 or
more non-colinear lat/lon points.