I changed the content model for dsgpolyo and dsgpolyx after our discussion today.  Since the fgdc standard calls for:
<xs:element name="dsgpolyo">
  <xs:complexType>
    <xs:choice>
      <xs:sequence>
        <xs:element ref="grngpoin"/>
        <xs:element ref="grngpoin"/>
        <xs:element ref="grngpoin"/>
        <xs:element ref="grngpoin" maxOccurs="unbounded"/>
      </xs:sequence>
      <xs:element ref="gring"/>
    </xs:choice>
  </xs:complexType>
</xs:element>

which seems incorrect (defining a polygon with at least 4 lat/lon points, rather
than 3 lat/lon points) (or a gring set), and since a polygon with 3 or 4 points that
are colinear don't define a very useful polygon, we decided to change the
content model to:

<xs:element name="dsgpolyo">
  <xs:complexType>
    <xs:choice>
      <xs:element ref="grngpoin" minOccurs="1" maxOccurs="unbounded"/>
      <xs:element ref="gring"/>
    </xs:choice>
  </xs:complexType>
</xs:element>


and would state in the documentation that polygons should be defined with 3 or
more non-colinear lat/lon points.

Also, i fixed a typo that we have been propogating: almax should be altmax, as
in maximum altitude according to the nbii standard.