Renamed elements in geologic time scale.
Made all four bounding box coordinates required, and explained the change in the
documentation to handle single point bounding areas.

Changed the content model to datasetGPolygonOuterRing to be a 3..n sequence.

This is a partial bug fix for bug 491: in that, structurally, the changes are
finished, but I still need to go through the documentation more closely.