Added new packaging changes to the eml, access, and party modules.
For eml, changed the "describes" attribute to be an element so that
it is repeatable.  Also, made the resources repeatable so that more
than one resource can be included in a package (eg eml can contain a
collection of ds:dataset and cit:citation).

For party, added the "id" attribute to Address and made it referencable.

For access, created a new complex type, and made it referenceable with an
ID.  Changed the "permission" enumeration to allow any string as well as the
enumerated values.