Checking in changes from dblankman for eml-constraint and eml-project.
See bug 486 for details on eml-constraint changes.  Note that I modified
David's schema slightly because the "ConstraintBase" element was not
needed in the hierarchy.

See bug 488 for eml-project change details.  This checkin is
exactly what David provided.

Both of these modules will need to be modified for the
new packaging scheme.