New model for handling the "content" side of the content/references choice.  Now, instead of having an inline "content" element that made for an ugly instance document, we now have the elements directly as children of the choice.  Before this was prevented because of the use of extended types.  Now, we switched ResourceBase to be a group called ResourceGroup, and EntityBase to be a group called EntityGroup, and used those groups in the content models rather than extending the types.
In going through the schemas, it is clear to me that we need to revisit all of the
nested schemas and determine which are required/optional.  I also eliminated
some duplicates of the "coverage" elements and others, and eliminated
some unneeded sequences.

Need to update all of the DTDs to use this same model, as well as the
documentation and formatting.

Also, I have not dealt with eml-constraint yet because I did not
understand the model as implemented and needed to talk it over
before making the changes.