Changed imports of major resource modules from refs to local elements cast as imported complex types.
Note: i did NOT remove the root elements from the corresponding resource modules. Im not sure if we should do this or not. In some IDE's it can be confusing to have both show up. On the other hand, in lower-level modules, having a root element defined means that one can have a free standing instance file of that module that can be independently validated.