Modified eml-coverage to change singleDateTime from a global element to a complex type named "SingleDateTimeType".  Now all singleDateTime elements are of this type, and changed the names of the rangeOfDates children to now be "beginDate" and "endDate", both of type SingleDateTimeType. This was agreed on our conference call on Friday.
NOTE: there are many global elements in eml-coverage.  I think this is going
to cause problems, because global elements must be namespace qualified (so
they'll require the namespace prefix to validate).  We need a test document
to test this, but I think we need to get rid of all of these global elements
in favor of complex types where reuse is needed, or put them inline if reuse
is not needed.