Merged the EML_DOCS_RC2_TEST branch back into the HEAD.  This branch changes:
1) The content model of the <doc:moduleDescription> tags to be of type eml-text.
This allows us to use some structure within the prose, and

2) Move the module descriptions out of the spec and into the .xsd files in order
to reduce the redundancy between the two.  Now, module descriptions should only
be hardcoded into the xsd (as it was originally), and the spec draws from that
text.

3) A few modules had leftover carriage returns, which I removed.

4) The spec document has a chapter that describes the logical use of the
modules.  This section is not yet complete.  I'm still working on the spatial
sections at the moment.