In eml-coverage.xsd:
1) Converted TemporalCoverage refs to complexTypes in order to be consistent
with the other xsd's.

2) made IdentifierName in TaxonomicCoverage of type rp:ResponsibleParty, by just
removing the originator element (which was redundant)

These are partial fixes to Bug 589

In looking at the documentation for eml-coverage, the <xs:attribute>s of id and
system consistently were not documented (as well in all the other xsd's.
However, the scope <xs:attribute> is drawn from res:ScopeType, and so is
documented in eml-resource.xsd.  Instead of documenting all of these repeating
attributes, I created two SimpleTypes in eml-resource, one IDType, the other
SystemType.  I changed each <xs:attribute name="id" ...> to be of type IDType.
Likewise for system.  Now, they work similarly to ScopeType, and are documented
in one location.  I used <xs:list itemType="xs:string /> to do this.

Lastly, I *didn't* make this change in eml-spatialVector.xsd yet, because there
are problems with it validating due to David's overwrite.  I don't think he's
fixed the problem yet, so I'm just going to wait to change that file so he
doesn't have to deal with even more conflicts.  I'll do it later.