Modifications to support the proposal in bug 596.  Removed attributeList from EntityGroup and instead added it into the content model for each of the types of entities (e.g., dataTable) so that it can be optional in otherEntity but required in the other entity types.  I simultaneously moved constraint as well because it doesn't make sense to have constraint be part of EntityGroup if attribute is not.
Changes occurred in entity, dataTable, spataialRaster, spatialVector,
storedProcedure, and view.

This change affects instance files by changing the order in which elements
occur (attributeList and constraint now follwo EntityGroup), but its a simple
change.  I updated all of the test instance files to deal with it.

Also, changed the id, system, and scope to use the IDType, SystemType and
ScopeType from eml-resource following Chris' model, so thay are all defined
in one place and thus need to be documented only once.