Made more changes to eml-access doumentation based on discussion with Matt. One of the changes is that following para was removed:
Note that, although access control may be bound to any element with an "id" attribute, the processing involved may be very costly.  For instance, it would not be recommended to apply access control to a column of a data file (eml-attribute), since every read/write operation on that column may not proceed until access is verified.

As we are only talking about distribution now, I dont think the above will be true now.