Removal of obsolete emlconversions code.
On 8/18/2006, Duane Costa wrote to eml-dev:

The eml module in CVS contains some old source code (eml/src/emlconversions) that was added by the LTER Network Office in 2003. This code is no longer actively developed or maintained. The application's purpose was to help the McMurdo LTER site convert their legacy metadata to EML, and this task has been accomplished. Also, LNO does retain a copy of this code in its own CVS. (Historically, I believe the code was added to the eml module because LNO at that time did not maintain its own CVS server.)

I would like to remove the code from the eml distribution, but Jing (wisely) suggested that I ping the eml-dev list in case there are any objections. If I hear no objections by the middle of next week I'll go ahead with the removal.