Add new instance field, dbFieldName, to store the name of an Attribute object's database field name. In many cases, the attribute name and the database field name will be equivalent, but there are two cases where they will differ:
(1) The attrbute name is not a legal database field name and must be modified to make it legal. (For example, by replacing space characters with underscore characters.)

(2) The attribute name is not unique within its attribute list, so the field name must be mangled to make it unique with the same database table.

Also added getter and setter methods for the dbFieldName field.