Commit it for callie. Made changes so that attributes will display vertical - changed the call template attributecommon to attributecommonvertical so that the template will display more compact view of the attribute information (vertical versus horizontal)
Used new class in the attributecommonvertical attribute types (highlightAttrib).
Changed attribute types columns so they will only display if there is content. Used xsl:when
Consolidated and changed names of some of the attribute type headings to save space.

Please note that this stylesheet could use some cleaning up, there are sections that are no longer applicable.
(horizontal view and shading of odd and even columns is not beings used). I think if someone wants to use the horizontal
view it might needs some editing. This xsl will need some cleaning up, perhaps when we make the next changes to the xsl
files.