Improved documentaiton on EML access module, making it clear that the access element belongs in the root eml node, not in the resource nodes.
Also fixed an error in the example with the references.  Before, the second
entity had an access element that referenced an entity123, which is wrong.
It should reference the access node itself, so I created an id for the
access node (access123), and pointed at this node in the references element.