Added `award` field.
Added first version of a new `award` field to the eml-project schema to add info that can be
used to reference funding information for the projet.  This provides the first implementation
of the `award` field, including the award title, identifier, and open funder registry identifier.
These proposed additions are partial implementations for issue #266.