Add a local copy of xml.xsd.
The version 2.1.1 internationalization changes introduced the need for
adding the xml.xsd file to support `xml:lang` attributes on
internationalized text strings.  However, the schema import pointed at a
network location of the schema file, which prevents validation when
offline.  Adding xml.xsd to the schema dir and referencing the local
copy allows offline validation to succeed and prevents injection attacks
on the remote schema file.  Need to verify that this is the right
version of xml.xsd.