Added support for new citation formats, fields, and inline citations.
These include a new ability to use Bibtex citation format both within
the `citation` element, and within a new `bibtex` element, to create
lists of refs using these in a literatureCited element (#300), as well
as in usageCitation (#259), and referencePublication (#277) elements.
All of this helps support data papers (#269), for which pandoc-style
citation keys can be used to cite these references in the text of
Markdown blocks in the EML document.  Added these features as
demonstrations in the eml-data-paper.xml sample document.