Update the deploy script.
THe script now builds the distribution site from scratch by running
the build_book.R script, and then copies source files like images and
schema docs and xsd files into the dist directory.  The dist directory
is then added to a temporary branch, and pushed to gh_pages to be picked
up by Netlify.  The temporary branch is then deleted, and we switch back
to the branch from which the original docs were built.